Fix some compiler warnings.
Again, given that char does not define whether it's signed or unsigned
in C, some compilers may issue warnings on these functions (leftover
from "various: Avoid using 'char' as an index" commit). At
power-supply.c, we cast to the right type for
sol_power_supply_get_capacity() usage.

+++ src/modules/flow/console/console.c
@@ -75,7 +75,7 @@ console_print_blob(struct console_data *mdata, const struct sol_blob *blob,
     buf = blob->mem;
     bufend = buf + blob->size;
     for (; buf < bufend; buf++) {
-        if (isprint(*buf))
+        if (isprint((uint8_t)*buf))
             fprintf(mdata->fp, "%#x(%c)", *buf, *buf);
         else
             fprintf(mdata->fp, "%#x", *buf);
